+++ week04/pepe/week04_pepe/week04_pepe/ContentArea.swift
@@ -79,6 +79,7 @@ struct ContentArea: View {
                     .foregroundColor(Color("ColorFontWhite"))
             }
         }
+        .background(Color("ColorBgBlack"))
     }
 }
