+++ week04/pepe/week04_pepe/week04_pepe/ToastView.swift
@@ -9,7 +9,22 @@ import SwiftUI
 
 struct ToastView: View {
     var body: some View {
-        Text("Hello, World!")
+        HStack {
+            Text("₩1,250,000부터")
+            Spacer()
+            Button {
+                EmptyView()
+            } label: {
+                Text("구입하기")
+                    .padding(.horizontal, 20)
+                    .padding(.vertical, 6)
+            }
+            .background(Color("ColorFontGray"))
+            .cornerRadius(17)
+        }
+        .padding(20)
+        .background(Color("ColorBtnSliver"))
+        .cornerRadius(17)
     }
 }
