+++ week04/pepe/week04_pepe/week04_pepe/week04_pepeApp.swift
@@ -11,7 +11,7 @@ import SwiftUI
 struct week04_pepeApp: App {
     var body: some Scene {
         WindowGroup {
-            DetectScrollPosition()
+            MainView()
         }
     }
 }
